Last Operator
- `Observable.last(defaultItem)` is used to retrieve the last item from the Observable and send it to the Observer as a `Single`.
When we retrieve an item using the `last` operator and it turns out the item is empty, the value in the parameter will become the default item.

- `Observable.lastElement()` is used to retrieve the last item from the Observable and send it to the Observer as `Maybe`.
If the Observable does not return a value, `Maybe` will be empty.
The difference with `last()` is that in the `lastElement()` operator we do not need to provide a defaultItem in the parameter.

- `Observable.lastOrError()` is similar to `last()`, but if the Observable does not return a value, it returns `NoSuchElementException`.