RxJava and RxAndroid Schedulers
A scheduler is a mechanism that determines in which thread an operation will be executed.

- Schedulers.io()
Supported by an unlimited thread pool, it is used for non-CPU-intensive I/O work such as network, database, and file system operations.
This scheduler is designed to perform I/O blocking operations asynchronously.

- Schedulers.computation()
Supported by a limited thread pool (up to the number of processors available), used for computational or CPU-intensive tasks such as data processing.
Be careful about allocating more compute threads than the number of cores available as this can cause performance degradation.

- Schedulers.newThread()
Creates a new thread for each task run, great for independent tasks.
This can be expensive because every time a task is executed, a new thread is created.

- Schedulers.from(Executor executor)
Creating a custom Scheduler using the existing executor Scheduler.from(Executors.newFixedThreadPool(n)), allows limiting the number of threads running simultaneously.

Schedulers.single()
Supported by a single thread that executes tasks sequentially based on the order of requests.

Schedulers.trampoline()
Runs a task FIFO in the same thread as the previously executed task. This is useful for maintaining the execution order of scheduled tasks.

AndroidSchedulers.mainThread()
Provided by RxAndroid to run tasks on the main thread or UI thread, used for UI updates.
Be careful not to overload this thread as it may cause ANRs.